Tomato-Web-App/Web-App using streamlit and VGG-19 Model/requirements.txt
numpy==1.23.5
opencv_python==4.6.0.66
Pillow==9.3.0
streamlit==1.15.1
tensorflow==2.11.0
tensorflow_intel==2.11.0
